Hotfix/revision android

+++ app/(tabs)/(settings)/accounts.tsx
@@ -15,7 +15,7 @@ import {
     getAmountOfTransactionsByAccountId, getTransactions,
     getTransactionsGroupedAndFiltered
 } from "@/lib/db";
-import {Account, FullTransaction, TransactionsGroupedByDate} from "@/lib/types/Transaction";
+import {Account, TransactionsGroupedByDate} from "@/lib/types/Transaction";
 import {useRouter} from "expo-router";
 import {changeEmoji} from "@/lib/store/features/ui/uiSlice";
 import * as ContextMenu from "zeego/context-menu";
@@ -29,10 +29,8 @@ import {
     updateTransactionsGroupedByCategory
 } from "@/lib/store/features/transactions/reportSlice";
 import {getCurrentMonth, getCurrentWeek} from "@/lib/helpers/date";
-import {selectCategories} from "@/lib/store/features/categories/categoriesSlice";
 import {useTranslation} from "react-i18next";
 import * as Haptics from "expo-haptics";
-import {onLongPress} from "@vueuse/core/index";
 import {useState} from "react";
 import OnlyDeleteOptionSheet from "@/lib/components/ui/android-dropdowns-sheets/OnlyDeleteOptionSheet";
 
@@ -54,7 +52,8 @@ export default function Screen() {
     const [selectedAccountId, setSelectAccountId] = useState<number>(0);
     const [open, setOpen] = useState<boolean>(false);
 
-    function onPressAccount(account: Account) {
+    async function onPressAccount(account: Account) {
+        await Haptics.selectionAsync();
         dispatch(updateAccountCreateUpdate(account));
         dispatch(changeEmoji(account.icon))
         router.push('/createEditAccount')
@@ -70,6 +69,11 @@ export default function Screen() {
                 text: t('COMMON.DELETE'),
                 isPreferred: false,
                 onPress: async () => {
+                    const allAccounts = getAllAccounts(db).length;
+                    if (allAccounts === 1) {
+                        Alert.alert(t('COMMON.WARNING'), t('COMMON.MESSAGES.MINIMUM_AMOUNT_ACCOUNTS'))
+                        return;
+                    }
                     await deleteAccount(db, accountId);
                     setSelectAccountId(0);
                     setOpen(false);
@@ -142,7 +146,7 @@ export default function Screen() {
                                                         color="$gray10Dark">{getAmountOfTransactionsByAccountId(db, account.id)} {t('COMMON.TRANSACTIONS')}</Text>
                                                 </YStack>
                                                 <Text
-                                                    fontSize={18}>{account.currency_symbol} {formatByThousands(account.balance.toString())} {account.currency_code}</Text>
+                                                    fontSize={18}>{account.currency_symbol} {formatByThousands(account.balance.toFixed(2))} {account.currency_code}</Text>
                                             </View>
                                         </TouchableOpacity>
                                     </ContextMenu.Trigger>
